Added --nodes to nimbusAdmin which shows which vm ids are running on which nodes moved helpconfig to RMIConfig

+++ tests/cc-nimbusAdminList-test.py
@@ -78,6 +78,26 @@ if rc != 0:
     print "error"
     sys.exit(1)
 
+cmd = "%s/bin/cloud-client.sh --run --name group --hours .25" % (cc_home)
+child = pexpect.spawn (cmd, timeout=to, maxread=20000, logfile=logfile)
+rc = child.expect ('Running:')
+if rc != 0:
+    print "group not found in the list"
+    sys.exit(1)
+handle = child.readline().strip().replace("'", "")
+rc = child.expect(pexpect.EOF)
+if rc != 0:
+    print "run"
+    sys.exit(1)
+
+cmd = "%s/bin/nimbus-admin --nodes" % (nimbus_home)
+print cmd
+(x, rc)=pexpect.run(cmd, withexitstatus=1)
+print x
+if rc != 0:
+    print "error"
+    sys.exit(1)
+
 cmd = "%s/bin/nimbus-admin --batch --shutdown --all" % (nimbus_home)
 print cmd
 (x, rc)=pexpect.run(cmd, withexitstatus=1)
